+++ python/dtype.h
@@ -5,9 +5,11 @@
 
 #include <string_view>
 
-#include "pybind11.h"
 #include <scipp/core/dtype.h>
 #include <scipp/units/unit.h>
+
+#include "pybind11.h"
+#include "unit.h"
 
 namespace pybind11 {
 class dtype;
@@ -19,7 +21,7 @@ scipp::core::DType scipp_dtype(const pybind11::object &type);
 
 std::tuple<scipp::core::DType, scipp::units::Unit>
 cast_dtype_and_unit(const pybind11::object &dtype,
-                    std::optional<scipp::units::Unit> unit);
+                    std::optional<ProtoUnit> unit);
 
 void ensure_conversion_possible(scipp::core::DType from, scipp::core::DType to,
                                 const std::string &data_name);