+++ python/src/scipp/plotting/resampling_model.py
@@ -2,6 +2,8 @@
 # Copyright (c) 2021 Scipp contributors (https://github.com/scipp)
 # @file
 # @author Simon Heybrock
+from enum import Enum
+
 from ..core import bin as bin_
 from ..core import dtype, units
 from ..core import linspace, rebin, get_slice_params, concatenate, histogram
@@ -9,12 +11,13 @@ from ..core import DataArray, DimensionError
 from .tools import to_bin_edges
 
 
-def _unit_requires_mean(obj):
-    return obj.unit != units.counts
+class ResamplingMode(Enum):
+    sum = 0
+    mean = 1
 
 
-def _resample(array, dim, edges):
-    if not _unit_requires_mean(array):
+def _resample(array, mode: ResamplingMode, dim, edges):
+    if mode == ResamplingMode.sum:
         return rebin(array, dim, edges)
     if array.dtype == dtype.float64:
         array = array.copy()
@@ -32,7 +35,7 @@ def _resample(array, dim, edges):
     array.data *= width
     unit = array.unit
     array.unit = units.counts
-    array = _resample(array, dim, edges)
+    array = rebin(array, dim, edges)
     width = edges[dim, 1:] - edges[dim, :-1]
     width.unit = units.one
     array /= width
@@ -41,7 +44,11 @@ def _resample(array, dim, edges):
 
 
 class ResamplingModel():
-    def __init__(self, array, resolution=None, bounds=None):
+    def __init__(self, array, *, resolution=None, bounds=None):
+        """
+        Model over a data array providing unified resampling functionality.
+        """
+        self._mode = None
         self._resolution = {} if resolution is None else resolution
         self._bounds = {} if bounds is None else bounds
         self._resampled = None
@@ -50,6 +57,15 @@ class ResamplingModel():
         self._home = None
         self._home_params = None
         self.update_array(array)
+
+    @property
+    def mode(self):
+        return self._mode
+
+    @mode.setter
+    def mode(self, m: ResamplingMode):
+        self._mode = ResamplingMode(m)
+        self.reset()
 
     @property
     def resolution(self):
@@ -80,13 +96,25 @@ class ResamplingModel():
     def edges(self):
         return self._edges
 
-    def _rebin(self, var, coords):
+    def _rebin(self, var, coords, *, sanitize=False):
+        """
+        Rebin a variable with given coords to new edges.
+
+        The `sanitize` parameter is an unfortunate artifact from mixing binned and
+        dense data in the case of masks of binned data. The called to _with_edges
+        below is done during setup of ResamplingDenseModel but not
+        ResamplingBinnedModel. A proper solution would likely include support for
+        non-edge inputs in `rebin`, or rather a function with similar constraits
+        such as `bin`, but for dense data.
+        """
         array = DataArray(data=var)
         for dim in coords:
             try:
                 array.coords[dim] = coords[dim]
             except DimensionError:  # Masks may be lower-dimensional
                 pass
+        if sanitize:
+            array, _ = _with_edges(array)
         plan = []
         for edge in self.edges:
             dim = edge.dims[-1]
@@ -99,14 +127,14 @@ class ResamplingModel():
                 plan.insert(0, edge)
         for edge in plan:
             try:
-                array = _resample(array, dim, edge)
+                array = _resample(array, self.mode, dim, edge)
             except RuntimeError:  # Limitation of rebin for slice of inner dim
-                array = _resample(array.copy(), edge.dims[-1], edge)
+                array = _resample(array.copy(), self.mode, edge.dims[-1], edge)
         return array
 
     def _make_edges(self, params):
         edges = []
-        for i, (dim, par) in enumerate(params.items()):
+        for dim, par in params.items():
             if isinstance(par, int):
                 continue
             low, high, unit, res = par
@@ -163,7 +191,9 @@ class ResamplingModel():
             self._edges = self._make_edges(params)
             self._resampled = self._resample(out)
             for name, mask in out.masks.items():
-                self._resampled.masks[name] = self._rebin(mask, out.meta).data
+                self._resampled.masks[name] = self._rebin(
+                    mask, out.meta, sanitize=isinstance(self,
+                                                        ResamplingBinnedModel)).data
             for dim in params:
                 size_one = self._resampled.sizes.get(dim, None) == 1
                 no_resolution = self.resolution.get(dim, None) is None
@@ -191,28 +221,45 @@ class ResamplingBinnedModel(ResamplingModel):
     def _make_array(self, array):
         return array
 
+    def _strip_masks(self, array):
+        array = array.copy(deep=False)
+        for name in list(array.masks.keys()):
+            del array.masks[name]
+        return array
+
     def _resample(self, array):
         # We could bin with all edges and then use `bins.sum()` but especially
         # for inputs with many bins handling the final edges using `histogram`
         # is faster with the current implementation of `sc.bin`.
-        edges = self.edges[-1]
-        dim = edges.dims[-1]
-        # TODO `bin` applies masks, but later we add rebinned masks. This is
+        # To avoid excessive memory use by `bin` we have to ensure that the
+        # dimension handled by `histogram` is not a dimension with many bins.
+        # We therefore select the smallest dimension of the input array for
+        # handling with `histogram`.
+        sizes = {}
+        for edges in self.edges:
+            dim = edges.dims[-1]
+            sizes[dim] = array.sizes[dim]
+        dim = min(sizes, key=sizes.get)
+        # `bin` applies masks, but later we add rebinned masks. This would be
         # inconsistent with how dense data is handled, where data is preserved
-        # even if masked, but masks grow. Note that manual masks handling for
-        # dense data in _call_resample is thus redundant, if it could be
-        # handled consistently here.
+        # even if masked, but masks grow. Therefore, we remove masks here. They
+        # get handled in _call_resample.
+        array = self._strip_masks(array)
         if dim in array.bins.coords:
+            index = list(sizes.keys()).index(dim)
+            edges = self.edges[index]
             # Must specify bounds for final dim despite handling by `histogram`
             # below: If coord is ragged binning would throw otherwise.
             bounds = concatenate(edges[dim, 0], edges[dim, -1], dim)
-            binned = bin_(array, edges=self.edges[:-1] + [bounds])
+            binned = bin_(
+                array,
+                edges=[bounds if i == index else e for i, e in enumerate(self.edges)])
             # TODO Use histogramming with "mean" mode once implemented
-            if _unit_requires_mean(binned.events):
+            if self.mode == ResamplingMode.mean:
                 return bin_(binned, edges=[edges]).bins.mean()
             else:
                 return histogram(binned, bins=edges)
-        elif _unit_requires_mean(array.events):
+        elif self.mode == ResamplingMode.mean:
             return bin_(array, edges=self.edges).bins.mean()
         else:
             return bin_(array, edges=self.edges).bins.sum()
@@ -222,6 +269,8 @@ def _with_edges(array):
     new_array = array.copy(deep=False)
     prefix = ''.join(array.dims)
     for dim, var in array.coords.items():
+        if dim not in array.dims:
+            continue
         new_array.coords[f'{prefix}_{dim}'] = var
         if var.sizes[dim] == array.sizes[dim]:
             new_array.coords[dim] = to_bin_edges(var, dim)