+++ python/tests/transform_coords_test.py
@@ -206,7 +206,7 @@ def test_dataset():
     assert sc.identical(transformed.coords['b'], a.rename_dims({'a': 'b'}))
 
 
-def test_binned():
+def make_binned():
     N = 50
     data = sc.DataArray(data=sc.ones(dims=['event'], unit=sc.units.counts, shape=[N]),
                         coords={
@@ -222,30 +222,67 @@ def test_binned():
     xbins = sc.Variable(dims=['x'], unit=sc.units.m, values=[0.1, 0.5, 0.9])
     ybins = sc.Variable(dims=['y'], unit=sc.units.m, values=[0.1, 0.5, 0.9])
     binned = sc.bin(data, edges=[xbins, ybins])
-    del binned.events.coords['y']
+    del binned.bins.coords['y']
     del binned.coords['y']
     binned.coords['y'] = sc.arange(dim='y', start=0, stop=2)
+    return binned
+
+
+def check_binned(da, original):
+    y = original.coords['y']
+    assert 'xy' not in original.bins.coords  # Buffer was copied
+    assert 'x' in original.bins.coords  # Buffer was copied for consume
+    assert sc.identical(da.bins.coords['xy'],
+                        (y * original.bins.coords['x']).rename_dims({'x': 'x2'}))
+    assert 'yy' not in da.bins.coords
+    assert sc.identical(da.coords['yy'], y * y)
+
+
+def test_binned():
+    binned = make_binned()
 
     def convert(*, x2, y):
         return {'yy': y * y, 'xy': x2 * y}
 
-    def check(da, original):
-        y = original.coords['y']
-        assert 'xy' not in original.events.coords  # Buffer was copied
-        assert 'x' in original.events.coords  # Buffer was copied for consume
-        assert sc.identical(da.bins.coords['xy'],
-                            (y * original.bins.coords['x']).rename_dims({'x': 'x2'}))
-        assert 'yy' not in da.events.coords
-        assert sc.identical(da.coords['yy'], y * y)
-
     graph = {'xy': convert, 'x2': 'x'}
     da = binned.transform_coords(['xy', 'yy'], graph=graph)
-    check(da, binned)
+    check_binned(da, binned)
     # If input is sliced, transform_coords has to copy the buffer
     da = binned['y', 0:1].transform_coords(['xy', 'yy'], graph=graph)
-    check(da, binned['y', 0:1])
+    check_binned(da, binned['y', 0:1])
     da = binned['y', 1:2].transform_coords(['xy', 'yy'], graph=graph)
-    check(da, binned['y', 1:2])
+    check_binned(da, binned['y', 1:2])
+
+
+def test_binned_no_dense_coord():
+    binned = make_binned()
+    del binned.coords['x']
+
+    def convert(*, x2, y):
+        return {'yy': y * y, 'xy': x2 * y}
+
+    graph = {('xy', 'yy'): convert, 'x2': 'x'}
+    da = binned.transform_coords(['xy', 'yy'], graph=graph)
+    check_binned(da, binned)
+    # works also without explicit request of 'yy' in graph
+    graph = {'xy': convert, 'x2': 'x'}
+    da = binned.transform_coords(['xy', 'yy'], graph=graph)
+    check_binned(da, binned)
+
+
+def test_binned_request_existing_consumed():
+    binned = make_binned()
+
+    def xy(*, x, y):
+        return x * y
+
+    def unused(x):
+        pass
+
+    graph = {'xy': xy, 'x': unused}
+    # Regression test, this used to throw due to an untested branch when
+    # requesting a previously consumed event coord
+    binned.transform_coords(['xy', 'x'], graph=graph)
 
 
 def test_cycles():